Add a unit test and move constant around
